Small itimer fixes and code reworking. Fixed a reentrant problem on __res_state
